fixing tests.
